Add a method to PeerManagerConnector to get connection ids
Adds a method to PeerManagerConnector to get a BiHashMap for
peer ids to connection ids. This will be used by PeerInterconnect
to get the peers and connection id that can be used for sending
messages.

To support this method, the PeerManager will handle a new message
PeerManagerRequest::ConnectionIds and the PeerMap has a new method
to return the BiHashMap.

This returns a BiHashMap because the PeerInterconnect will need to
go convert the ids to each other, where the direction depends on
if the message is incoming or outgoing.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>