Implement initial PeerInterconnect
PeerInterconnect will receive incoming messages from peers and
dispatch them to the NetworkMessageType handlers. It will also receive
messages from handlers that need to be sent to other peers.

When an incoming message is received, the connection_id is converted to
a peer_id. The reverse is done for an outgoing message. If a message is
received from an unknown connection, the PeerInterconnect will request
the current peers from the PeerManager and update the local map
of peers.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>