Add v1 Trust Peer Authorization
This commit includes the message handlers for the new authorization
message, as well as updating the authorization state machine for
new actions/states for the new trust procedure.

The v0 Trust is still supported along with v1. If a ConnectRequest
is the first message receive v0 trust will be used. If
AuthProtocolRequest is the first message v1 Trust will be used.
This is required to stay backwards compatible with 0.4.

It is also now required to track both out local authorization state
and the state of the remote as separate state machines. This is
because authorization will be happening in parallel.

The new implementation is behind a new experimental features
`trust-authorization`. The `challenge-authorization` features
now depends on this new feature.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>