Update AuthProtocolRequest/Response to handle required authorization
If we know what the required authorization type is, the protocol request
should only respond with that authorization type even if other auth
types are configured.

When handling the response, the node should also only pick the correct
required authorization type. In both cases if the required authorization
type is not offered or not configured, and AuthorizationError will be
returned.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>