Wrap the SigningContext in an Arc<Mutex>
This is required because the authorization manager will require
a VerifierFactory. Currently the factory is the context but the
context was being passed to the scabbard factory. It is not
recommended to create more then one context so it has to be
wrapped in an arc mutex so it can be shared.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>