Add configuration for which key to use for peering with --peer
What the local authorization will be for unidentified peers must be
known at time of peering to properly close duplicate connections.
This authorization should default to challenge if challenge can
be enabled, however there may be more then one key configured.

--peering-key allows a user to configure which key is used for
challenge authorization. Default to the key named splinterd.
If only one key is configured for use in challenge auth and that
is not called splinterd, that key will be set instead.

If no keys are provided, the token defaults to trust.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>