Log a warning instead of returning an error on_peer_connected
Before if a peer connected and token_to_peer does not have the
peer an error would be returned. However, this should not be
treated as an error because this is expected if the peer that
connected started as a unidentified peer. Instead, just log
a warning and do not send a protocol request.

A protocol request will be sent when a circuit request is
submitted.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>